Refactor: 시작점 선택 페이지 명명 변경 및 컴포넌트 명명 변경
