--- src/components/startPointSelectionPage/CarouselPage/index.jsx
@@ -1,48 +1,0 @@
-import { useState } from "react";
-import { FaCircleChevronLeft, FaCircleChevronRight } from "react-icons/fa6";
-
-import useUrlStore from "../../../store";
-import SelectionPage from "../SelectionPage";
-
-function CarouselContainer() {
-  const { videoUrls, audioUrls } = useUrlStore();
-  const [currentIndex, setCurrentIndex] = useState(0);
-
-  // TO DO: store 수정 필요
-  const videoUrlList = Object.values(videoUrls);
-  const audioUrlList = Object.values(audioUrls);
-
-  const CAROUSEL_ICON_SIZE = 40;
-
-  function handleLeftClick() {
-    setCurrentIndex((prev) => prev - 1);
-  }
-
-  function handleRightClick() {
-    setCurrentIndex((prev) => prev + 1);
-  }
-
-  return (
-    <main className="flex flex-col items-center justify-center w-full h-full">
-      <section className="flex items-center space-x-30">
-        <FaCircleChevronLeft
-          size={CAROUSEL_ICON_SIZE}
-          className={`fill-white ${currentIndex === 0 ? "invisible" : null} cursor-pointer`}
-          onClick={handleLeftClick}
-        />
-        <SelectionPage
-          isMain={currentIndex === 0}
-          selectedVideoUrl={videoUrlList[currentIndex]}
-          selectedAudioUrl={audioUrlList[currentIndex]}
-        />
-        <FaCircleChevronRight
-          size={CAROUSEL_ICON_SIZE}
-          className={`fill-white ${currentIndex === 2 ? "invisible" : null} cursor-pointer`}
-          onClick={handleRightClick}
-        />
-      </section>
-    </main>
-  );
-}
-
-export default CarouselContainer;
